+++ Battery/BatteryState.swift
@@ -1,0 +1,13 @@
+//
+//  BatteryState.swift
+//  Battery
+//
+//  Created by Gabriel Soria Souza on 13/07/2022.
+//
+
+enum BatteryState: String {
+    case unknown
+    case charging
+    case unplugged
+    case full
+}
